move(OrderTable): EatInOrder domain패키지 가독성을 위해 OrderTable 관련 패키지 분리
