+++ src/main/kotlin/kitchenpos/order/tobe/eatinorder/application/OrderTableService.kt
@@ -2,11 +2,11 @@ package kitchenpos.order.tobe.eatinorder.application
 
 import java.util.*
 import kitchenpos.order.tobe.eatinorder.application.dto.CreateOrderTableReq
-import kitchenpos.order.tobe.eatinorder.domain.OrderTable
-import kitchenpos.order.tobe.eatinorder.domain.OrderTableEmptyService
-import kitchenpos.order.tobe.eatinorder.domain.OrderTableName
-import kitchenpos.order.tobe.eatinorder.domain.OrderTableOccupancy
-import kitchenpos.order.tobe.eatinorder.domain.OrderTableRepository
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTable
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTableEmptyService
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTableName
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTableOccupancy
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTableRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
