--- src/main/kotlin/kitchenpos/order/tobe/eatinorder/domain/OrderTableStatus.kt
@@ -1,5 +1,0 @@
-package kitchenpos.order.tobe.eatinorder.domain
-
-enum class OrderTableStatus {
-    EMPTY, OCCUPIED,
-}
