+++ src/main/kotlin/kitchenpos/order/tobe/eatinorder/domain/ordertable/OrderTableEmptyService.kt
@@ -1,4 +1,4 @@
-package kitchenpos.order.tobe.eatinorder.domain
+package kitchenpos.order.tobe.eatinorder.domain.ordertable
 
 interface OrderTableEmptyService {
     fun canEmpty(orderTable: OrderTable): Boolean
