+++ src/main/kotlin/kitchenpos/order/tobe/eatinorder/domain/ordertable/OrderTableStatus.kt
@@ -1,0 +1,5 @@
+package kitchenpos.order.tobe.eatinorder.domain.ordertable
+
+enum class OrderTableStatus {
+    EMPTY, OCCUPIED,
+}
