+++ src/main/kotlin/kitchenpos/order/tobe/eatinorder/infra/DefaultOrderTableEmptyService.kt
@@ -3,8 +3,8 @@ package kitchenpos.order.tobe.eatinorder.infra
 import kitchenpos.common.annotation.DomainService
 import kitchenpos.order.tobe.eatinorder.domain.EatInOrderRepository
 import kitchenpos.order.tobe.eatinorder.domain.EatInOrderStatus
-import kitchenpos.order.tobe.eatinorder.domain.OrderTable
-import kitchenpos.order.tobe.eatinorder.domain.OrderTableEmptyService
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTable
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTableEmptyService
 
 @DomainService
 class DefaultOrderTableEmptyService(
