+++ src/test/kotlin/kitchenpos/order/tobe/eatinorder/infra/FakeOrderTableRepository.kt
@@ -1,8 +1,8 @@
 package kitchenpos.order.tobe.eatinorder.infra
 
 import java.util.*
-import kitchenpos.order.tobe.eatinorder.domain.OrderTable
-import kitchenpos.order.tobe.eatinorder.domain.OrderTableRepository
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTable
+import kitchenpos.order.tobe.eatinorder.domain.ordertable.OrderTableRepository
 
 
 class FakeOrderTableRepository(
